Renamed packages and add return statement
* Added return statement to return this class for JenkinsXmlData
  this enables re use of class, example:
  jenkinsData.addParameter(parram1).addParameter(param2);
