+++ pom.xml
@@ -1,9 +1,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>eiffel-community</groupId>
-    <artifactId>eiffel-shared-java</artifactId>
+    <artifactId>eiffel-commons-java</artifactId>
     <version>0.0.1</version>
-    <name>eiffel shared java</name>
+    <name>eiffel commons java</name>
     <description>A shared library for eiffel components</description>
 
     <properties>
