+++ src/main/java/com/ericsson/ec/helpers/MediaType.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.esj.helpers;
+package com.ericsson.ec.helpers;
 
 public class MediaType {
