+++ src/main/java/com/ericsson/ec/utils/HttpRequest.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package utils;
+package com.ericsson.ec.utils;
 
 import java.io.File;
 import java.io.IOException;
